refactor: consolidate route-level gates into auth::guard
require_authenticated, require_role, and require_scope now live in a
single module that owns identity-presence/role/scope gating regardless
of authentication source. Resolves the original smell: require_authenticated
previously lived in rbac because no other module owned 'authenticated user'.
