+++ src/auth/role/mod.rs
@@ -10,8 +10,9 @@
 //! - [`RoleExtractor`] — trait to resolve the current user's role from a request.
 //! - [`middleware()`] — Tower layer that calls your extractor and stores [`Role`] in extensions.
 //! - [`Role`] — newtype extractor over `String`; pull the resolved role into handlers.
-//! - [`require_role()`] — guard layer that rejects requests whose role is not in the allowed list.
-//! - [`require_authenticated()`] — guard layer that rejects requests with no role at all.
+//!
+//! Route-level gating (`require_role`, `require_authenticated`) lives in
+//! [`crate::auth::guard`].
 //!
 //! # Wiring order
 //!
@@ -21,7 +22,7 @@
 //!
 //! ```rust,no_run
 //! use axum::{Router, routing::get};
-//! use modo::auth::role::{self, RoleExtractor};
+//! use modo::auth::{guard, role::{self, RoleExtractor}};
 //! use modo::Result;
 //!
 //! struct MyExtractor;
@@ -34,16 +35,14 @@
 //!
 //! let app: Router = Router::new()
 //!     .route("/admin", get(|| async { "ok" }))
-//!     .route_layer(role::require_role(["admin", "owner"]))
+//!     .route_layer(guard::require_role(["admin", "owner"]))
 //!     .layer(role::middleware(MyExtractor));
 //! ```
 
 mod extractor;
-mod guard;
 mod middleware;
 mod traits;
 
 pub use extractor::Role;
-pub use guard::{require_authenticated, require_role};
 pub use middleware::middleware;
 pub use traits::RoleExtractor;
