+++ tests/apikey_test.rs
@@ -271,7 +271,8 @@ use axum::body::Body;
 use http::{Request, Response, StatusCode};
 use tower::{Layer, ServiceExt};
 
-use modo::auth::apikey::{ApiKeyLayer, ApiKeyMeta, require_scope};
+use modo::auth::apikey::{ApiKeyLayer, ApiKeyMeta};
+use modo::auth::guard::require_scope;
 
 /// Inner service that reads ApiKeyMeta from extensions and echoes the tenant_id.
 async fn echo_handler(req: Request<Body>) -> Result<Response<Body>, Infallible> {
